+++ app/src/main/java/com/cajetan/youtubeplus/PlayerActivity.java
@@ -13,8 +13,12 @@ import android.support.v4.media.session.MediaSessionCompat;
 import android.support.v4.media.session.PlaybackStateCompat;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import android.util.Log;
+import android.widget.Toast;
 
 import com.cajetan.youtubeplus.utils.FullScreenHelper;
+import com.cajetan.youtubeplus.utils.YouTubeData;
+import com.google.api.services.youtube.model.Video;
 import com.pierfrancescosoffritti.androidyoutubeplayer.player.PlayerConstants;
 import com.pierfrancescosoffritti.androidyoutubeplayer.player.YouTubePlayer;
 import com.pierfrancescosoffritti.androidyoutubeplayer.player.YouTubePlayerView;
@@ -23,7 +27,7 @@ import com.pierfrancescosoffritti.androidyoutubeplayer.player.listeners.YouTubeP
 import com.pierfrancescosoffritti.androidyoutubeplayer.player.listeners.YouTubePlayerInitListener;
 import com.pierfrancescosoffritti.androidyoutubeplayer.utils.YouTubePlayerTracker;
 
-public class PlayerActivity extends AppCompatActivity {
+public class PlayerActivity extends AppCompatActivity implements YouTubeData.VideoDataListener {
     private static final String TAG = PlayerActivity.class.getSimpleName();
 
     private YouTubePlayerView mainPlayerView;
@@ -32,16 +36,43 @@ public class PlayerActivity extends AppCompatActivity {
     private static MediaSessionCompat mMediaSession;
     private PlaybackStateCompat.Builder mStateBuilder;
 
+    private YouTubeData youTubeData;
+    private Video mVideoData;
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_player);
+
+        // TODO: wtf
+        String videoUrl = null;
+        if (getIntent() != null && getIntent().getExtras() != null)
+            videoUrl = getIntent().getExtras().getString(Intent.EXTRA_TEXT);
+
+        String videoId;
+
+        // Activity started by a regular Intent with a video id
+        if (getIntent().getExtras().containsKey(getString(R.string.video_id_key)))
+            videoId = getIntent().getExtras().getString(getString(R.string.video_id_key));
+         // Activity started by a share Intent with a video url
+        else if (videoUrl != null && !videoUrl.equals(""))
+            videoId = videoUrl.substring(videoUrl.length()-11, videoUrl.length());
+        // No video to play, throw an exception
+        else {
+            // TODO: throw something more informative here
+            throw new IllegalArgumentException("...");
+        }
+
+        youTubeData = new YouTubeData(this);
+        youTubeData.getVideoDataById(videoId);
 
         mainPlayerView = findViewById(R.id.main_player_view);
 
         setupPlayer();
         setupMediaSession();
     }
+
+
 
     private void setupPlayer() {
         // TODO: maybe refactor a bit
@@ -92,6 +123,7 @@ public class PlayerActivity extends AppCompatActivity {
 
                         mMediaSession.setPlaybackState(mStateBuilder.build());
                         showMediaNotification(mStateBuilder.build());
+
                         super.onStateChange(state);
                     }
                 });
@@ -207,9 +239,10 @@ public class PlayerActivity extends AppCompatActivity {
     }
 
     @Override
-    protected void onDestroy() {
-        mainPlayerView.release();
-        super.onDestroy();
+    public void onVideoDataReceived(Video videoData) {
+        mVideoData = videoData;
+        Toast.makeText(this, mVideoData.getSnippet().getTitle(), Toast.LENGTH_LONG).show();
+        showMediaNotification(mStateBuilder.build());
     }
 
     // TODO: test with and without singletop
@@ -218,4 +251,16 @@ public class PlayerActivity extends AppCompatActivity {
 //    protected void onNewIntent(Intent intent) {
 //        super.onNewIntent(intent);
 //    }
+
+    @Override
+    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
+        youTubeData.onParentActivityResult(requestCode, resultCode, data);
+        super.onActivityResult(requestCode, resultCode, data);
+    }
+
+    @Override
+    protected void onDestroy() {
+        mainPlayerView.release();
+        super.onDestroy();
+    }
 }