+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ---
 
+#### 11.11.2018
+
+* [General] Added a Toast with the video title 
+* [General - Dev] Managed to get data from the API
+
+#### 10.11.2018
+
+* [General - Dev] Started implementing YouTubeData API call
+
 #### 8.11.2018
 
 * [General] Gradle versioning system